Separated boards to their own respective controller and routes files
