+++ server/routes/boards.js
@@ -1,0 +1,6 @@
+const router = require("express").Router();
+const boardsController = require("../controllers/boards");
+
+router.get("/", boardsController.getBoard);
+
+module.exports = router;
